Implemented Register function
I implemented the register in the backend and in the frontend. I implemented the funcitonality in the backend to allow the creation of a user by the name of a register post request and the frontend send the information to the backend with the new page created for registration. Tested and it works